Nov5-2022

+++ CPTool/NewPages/Components/SimpleListComponent.razor.cs
@@ -10,15 +10,14 @@ namespace CPTool.NewPages.Components
 
         where TGedById : GetByIdQuery, new()
     {
-        [Inject]
-        public IMediator mediator { get; set; }
+        
         TMasterList ModelList { get; set; } = new();
         [Parameter]
 
         public List<TMaster> ElementsMasters { get; set; } = new();
-
         [Parameter]
-
+        public RenderFragment MasterOtherButtons { get; set; }
+        [Parameter]
         public EventCallback<List<TMaster>> ElementsMastersChanged { get; set; }
         [Parameter]
         public TMaster SelectedMaster { get; set; } = new();
@@ -33,73 +32,86 @@ namespace CPTool.NewPages.Components
         [Parameter]
         [EditorRequired]
         public Func<TMaster, Task<DialogResult>> OnShowDialogMaster { get; set; }
-
+        [Parameter]
+        public EventCallback OnAdd { get; set; }
+        [Parameter]
+        public EventCallback OnEdit { get; set; }
+        [Parameter]
+        public EventCallback OnDelete { get; set; }
         async Task RowClickedMaster(TableRowClickEventArgs<TMaster> eq)
         {
             if (eq.Item.Id == SelectedMaster.Id) return;
-            TGedById getbyid = new()
-            {
-                Id = eq.Item.Id,
-            };
-            SelectedMaster = await mediator.Send(getbyid) as TMaster;
+           
+            SelectedMaster = eq.Item;
             await SelectedMasterChanged.InvokeAsync(SelectedMaster);
 
         }
 
         async Task Add()
         {
-            TMaster model = new();
-
-            var result = await OnShowDialogMaster.Invoke(model);
-            if (!result.Cancelled)
+            if (OnAdd.HasDelegate)  await OnAdd.InvokeAsync();
+            else
             {
-                if(result.Data is TMaster)
+                TMaster model = new();
+
+                var result = await OnShowDialogMaster.Invoke(model);
+                if (!result.Cancelled)
                 {
-                    var data = result.Data as TMaster;
-                    TGedById querydetail = new() { Id = data.Id };
-                    SelectedMaster = await mediator.Send(querydetail) as TMaster;
+                    if (result.Data is TMaster)
+                    {
+                        var data = result.Data as TMaster;
+                        TGedById querydetail = new() { Id = data.Id };
+                        SelectedMaster = await Mediator.Send(querydetail) as TMaster;
 
 
-                    ElementsMasters = (await mediator.Send(ModelList)) as List<TMaster>;
-                    await ElementsMastersChanged.InvokeAsync(ElementsMasters);
-                    await SelectedMasterChanged.InvokeAsync(SelectedMaster);
+                        ElementsMasters = (await Mediator.Send(ModelList)) as List<TMaster>;
+                        await ElementsMastersChanged.InvokeAsync(ElementsMasters);
+                        await SelectedMasterChanged.InvokeAsync(SelectedMaster);
+                    }
+
                 }
-                
             }
+            
         }
         async Task Edit()
         {
-
-            var result = await OnShowDialogMaster.Invoke(SelectedMaster);
-
-            if (!result.Cancelled)
+            if (OnEdit.HasDelegate) await OnEdit.InvokeAsync();
+            else
             {
-                if (result.Data is TMaster)
+                var result = await OnShowDialogMaster.Invoke(SelectedMaster);
+
+                if (!result.Cancelled)
                 {
-                    var data = result.Data as TMaster;
-                    TGedById querydetail = new() { Id = data.Id };
-                    SelectedMaster = await mediator.Send(querydetail) as TMaster;
-                    ElementsMasters = (await mediator.Send(ModelList)) as List<TMaster>;
-                    await ElementsMastersChanged.InvokeAsync(ElementsMasters);
-                    await SelectedMasterChanged.InvokeAsync(SelectedMaster);
+                    if (result.Data is TMaster)
+                    {
+                        var data = result.Data as TMaster;
+                        TGedById querydetail = new() { Id = data.Id };
+                        SelectedMaster = await Mediator.Send(querydetail) as TMaster;
+                        ElementsMasters = (await Mediator.Send(ModelList)) as List<TMaster>;
+                        await ElementsMastersChanged.InvokeAsync(ElementsMasters);
+                        await SelectedMasterChanged.InvokeAsync(SelectedMaster);
+                    }
                 }
             }
         }
         async Task Delete()
         {
-            TDeleteMaster ModelDelete = new() { Id = SelectedMaster.Id, Name = SelectedMaster.Name };
-
-            var result = await ToolDialogService.DeleteRow(ModelDelete);
-
-            if (!result.Cancelled)
+            if (OnDelete.HasDelegate) await OnDelete.InvokeAsync();
+            else
             {
-                SelectedMaster = new();
-                ElementsMasters = (await mediator.Send(ModelList)) as List<TMaster>;
-                await ElementsMastersChanged.InvokeAsync(ElementsMasters);
-                await SelectedMasterChanged.InvokeAsync(SelectedMaster);
+                TDeleteMaster ModelDelete = new() { Id = SelectedMaster.Id, Name = SelectedMaster.Name };
 
+                var result = await ToolDialogService.DeleteRow(ModelDelete);
+
+                if (!result.Cancelled)
+                {
+                    SelectedMaster = new();
+                    ElementsMasters = (await Mediator.Send(ModelList)) as List<TMaster>;
+                    await ElementsMastersChanged.InvokeAsync(ElementsMasters);
+                    await SelectedMasterChanged.InvokeAsync(SelectedMaster);
+
+                }
             }
-
 
         }
